Quick Script to Split huge file to smaller files
This script splits the HUGE file, 22 Million rows to 1 Million rows files, and it does it pretty quick.
It doesn't store the huge file in memory, rather it reads rows to buffer and writes them to small file until it reaches 1 Million rows. 
Usage in cmd: 
python split.py all_sessions.csv
The result in my case is 23 "chunks" each is ~340MB except the last one which is 176MB.
Be aware that the Header row appears only in the first chunk, this can be easily modified.
Also, we can modify this script to create sample files instead of using the whole rows or columns.